No need to include the Chess.app source or the Sjeng engine.
I was including a tarball of the Chess.app source because I had copied
some of its code and wasn't sure if that was a licensing requirement.
But I'm no longer using any code from Chess.app -- ChessFidget is
100% Swift.

I was including a binary executable of the Sjeng engine on the theory
that I wouldn't have to rely on Chess.app being in /Applications.  But
I think it's safe to assume it'll be there.